frontend/requirements.txt
questionary
rich
streamlit
fastapi
uvicorn
firebase-admin
pyrebase4
